fix: tasks 3,4,6,7 for theme 'exception handling' (separate demo and methods)
