+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task3.java
@@ -4,19 +4,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class Task3 {
-    public static void main(String[] args) {
-        String name = "*Anton_123";
-
-        try {
-           if (isNameCorrect(name)) {
-               System.out.println("Welcome, " + name);
-           }
-        } catch (ImpossibleAccountNameException forbidden) {
-            System.out.println(forbidden.getMessage());
-            forbidden.printStackTrace(System.out);
-        }
-    }
-
     public static boolean isNameCorrect (String string) throws ImpossibleAccountNameException {
         Pattern pattern = Pattern.compile("\\W");
         Matcher matcher = pattern.matcher(string);
