+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task5.java
@@ -3,7 +3,6 @@ package com.rakovets.course.java.core.practice.exception_handling;
 import java.util.Locale;
 
 public class Task5 {
-
     public static String getCorrectName (String string) throws ImpossibleAccountNameException {
         try {
             string = string.toUpperCase(Locale.ROOT).trim() + "123";
