+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/Task6.java
@@ -4,16 +4,6 @@ import java.util.NoSuchElementException;
 import java.util.Random;
 
 public class Task6 {
-    public static void main(String[] args) {
-        try {
-            getException();
-        } catch (NullPointerException | NoSuchElementException ex) {
-            ex.printStackTrace(System.out);
-        } catch (ArrayIndexOutOfBoundsException arx) {
-            System.out.println("ArrayIndexOutOfBoundsException: " + arx.getMessage());
-        }
-    }
-
     public static void getException() throws NullPointerException, NoSuchElementException, ArrayIndexOutOfBoundsException {
         Random random = new Random();
         int result = random.nextInt(10);
